# backend/src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Homni Togli API
  version: '2.0.0'
  description: Per-project feature toggle management with RBAC and API key authentication.
servers:
  - url: 'http://localhost:8080'
    description: local
security:
  - apiKey: []
  - oidc: []
...
paths:
...
  /projects/{projectId}/toggles:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Feature Toggles]
      operationId: listToggles
      summary: List feature toggles in a project
      description: |
        Requires READ_TOGGLES permission (READER+). Filter semantics:
        - `enabled` + `environment` → toggles whose state in that specific env matches
        - `enabled` only → toggles enabled (or disabled) in at least one env
        - `environment` only → toggles assigned to that env regardless of state
      parameters:
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
        - name: environment
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleListResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'

    post:
      tags: [Feature Toggles]
      operationId: createToggle
      summary: Create a new feature toggle
      description: Requires WRITE_TOGGLES permission (EDITOR+). Project must not be archived.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureToggleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'
  /projects/{projectId}/toggles/{toggleId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: toggleId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Feature Toggles]
      operationId: getToggle
      summary: Get toggle by ID
      description: Requires READ_TOGGLES permission (READER+).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleSingleResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    patch:
      tags: [Feature Toggles]
      operationId: updateToggle
      summary: Update a feature toggle
      description: Partial update. Requires WRITE_TOGGLES permission (EDITOR+).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFeatureToggleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureToggleSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    delete:
      tags: [Feature Toggles]
      operationId: deleteToggle
      summary: Delete a feature toggle
      description: Requires WRITE_TOGGLES permission (EDITOR+).
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

  # ========== Environments ==========
...
  /projects/{projectId}/members:
    parameters:
      - $ref: '#/components/parameters/projectId'

    get:
      tags: [Members]
      operationId: listMembers
      summary: List project members
      description: Requires MANAGE_MEMBERS permission (ADMIN).
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipListResponse'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
  /projects/{projectId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    put:
      tags: [Members]
      operationId: upsertMember
      summary: Add or update a project member
      description: Adds the user to the project with the specified role, or updates the role if already a member. Requires MANAGE_MEMBERS permission (ADMIN).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeMemberRoleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipSingleResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    delete:
      tags: [Members]
      operationId: removeMember
      summary: Remove a member from the project
      description: Requires MANAGE_MEMBERS permission (ADMIN). Project must not be archived.
      responses:
        '204':
          $ref: '#/components/responses/204'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

  # ========== Users ==========
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    '204':
      description: No content
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ResponseMeta:
      type: object
      required: [timestamp]
      properties:
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- Project ----------
...
    CreateFeatureToggleRequest:
      type: object
      required: [name, environments]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: dark-mode
        description:
          type: string
          maxLength: 1000
        environments:
          type: array
          items:
            type: string
          example: [DEV, PROD]
    UpdateFeatureToggleRequest:
      type: object
      description: |
        Partial update for a feature toggle. Any field can be omitted. The
        `environments` array (if present) replaces the toggle's assigned env
        set; envs added this way start disabled, removed envs lose their state.
        The `environmentStates` map (if present) flips the enabled flag for
        each listed env independently — pass a one-entry map for an inline
        switch click, a multi-entry map for bulk form-style edits.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        environments:
          type: array
          items:
            type: string
        environmentStates:
          type: object
          description: Map of environment name → desired enabled flag.
          additionalProperties:
            type: boolean
          example:
            DEV: true
            PROD: false
    FeatureToggleEnvironment:
      type: object
      required: [name, enabled]
      properties:
        name:
          type: string
        enabled:
          type: boolean
    FeatureToggle:
      type: object
      required: [id, projectId, name, environments, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        environments:
          type: array
          description: |
            Per-environment state of the toggle. Each entry indicates that the
            toggle is "present" in that env, plus the independent enabled flag
            for that env.
          items:
            $ref: '#/components/schemas/FeatureToggleEnvironment'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    FeatureToggleSingleResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          $ref: '#/components/schemas/FeatureToggle'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    FeatureToggleListResponse:
      type: object
      required: [payload, pagination, meta]
      properties:
        payload:
          type: array
          items:
            $ref: '#/components/schemas/FeatureToggle'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- Environment ----------
...
    ChangeMemberRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [ADMIN, EDITOR, READER]
    ProjectMembership:
      type: object
      required: [id, projectId, userId, role, grantedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [ADMIN, EDITOR, READER]
        email:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        grantedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    MembershipSingleResponse:
      type: object
      required: [payload, meta]
      properties:
        payload:
          $ref: '#/components/schemas/ProjectMembership'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    MembershipListResponse:
      type: object
      required: [payload, pagination, meta]
      properties:
        payload:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMembership'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # ---------- User ----------
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    oidc:
      type: http
      scheme: bearer
      bearerFormat: JWT